Remove unneeded(?) case statement
This seems to be legacy code that's kept because it's legacy and no one has bothered to look into it. If stuff breaks I'll fix it, but having this stupid whitelist of filesystems is unneeded, and I can't think of a single reason to use `shred`, especially since so many of use use SSD:s (where `shred` does the opposite of good).